rename issue file to issues for naming consistency
