+++ src/applications/appeals/10182/pages/contestableIssues.js
@@ -6,7 +6,7 @@ import { maxIssues } from '../validations/issues';
 
 import { hasSomeSelected } from '../../shared/utils/issues';
 import { SELECTED } from '../../shared/constants';
-import { selectionRequired } from '../../shared/validations/issue';
+import { selectionRequired } from '../../shared/validations/issues';
 
 /**
  * contestable issues with add issue link (list loop)
