+++ src/applications/appeals/10182/tests/validations/issues.unit.spec.js
@@ -10,7 +10,7 @@ import {
 
 import { MAX_LENGTH, SELECTED } from '../../../shared/constants';
 import { getDate } from '../../../shared/utils/dates';
-import { selectionRequired } from '../../../shared/validations/issue';
+import { selectionRequired } from '../../../shared/validations/issues';
 
 const _ = null;
