+++ src/applications/appeals/10182/validations/issues.js
@@ -5,7 +5,7 @@ import { content as extensionReasonContent } from '../content/extensionReason';
 import { maxSelectedErrorMessage } from '../../shared/content/contestableIssues';
 import { MAX_LENGTH, REGEXP } from '../../shared/constants';
 import { getSelected } from '../../shared/utils/issues';
-import { addUniqueIssueErrorMessage } from '../../shared/validations/issue';
+import { addUniqueIssueErrorMessage } from '../../shared/validations/issues';
 
 // Alert Veteran to duplicates based on name & decision date
 export const uniqueIssue = (
