+++ src/applications/appeals/996/pages/contestableIssues.js
@@ -4,7 +4,7 @@ import { maxIssues } from '../validations/issues';
 
 import { SELECTED } from '../../shared/constants';
 import { hasSomeSelected } from '../../shared/utils/issues';
-import { selectionRequired } from '../../shared/validations/issue';
+import { selectionRequired } from '../../shared/validations/issues';
 
 /**
  * contestable issues with add issue link (list loop)
