+++ src/applications/appeals/testing/components/AddIssue.jsx
@@ -30,7 +30,7 @@ import {
   SELECTED,
 } from '../../shared/constants';
 import { getSelected, calculateIndexOffset } from '../../shared/utils/issues';
-import { checkValidations } from '../../shared/validations/issue';
+import { checkValidations } from '../../shared/validations/issues';
 
 const ISSUES_PAGE = `/${CONTESTABLE_ISSUES_PATH}`;
